% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry-data.R
\docType{data}
\name{berry_lookup}
\alias{berry_lookup}
\title{Data containing berry names and IDs. Contains all those berries that are
available on the PokeAPI database.}
\format{An object of class \code{"data.frame"}.}
\usage{
data(berry_lookup)
}
\description{
Data containing berry names and IDs. Contains all those berries that are
available on the PokeAPI database.
}
\keyword{datasets}
