% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution_chain.R
\docType{data}
\name{evolution_chain_ids}
\alias{evolution_chain_ids}
\title{Evolution chain IDs of pokemon characters on pokeapi V2.}
\format{A data frame with 415 rows and 2 variables:
\describe{
  \item{id}{id, used for lookup in API}
  \item{name}{name of pokemon character}
}}
\usage{
data(evolution_chain)
}
\description{
Evolution chain IDs of pokemon characters on pokeapi V2.
}
\keyword{datasets}
