% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution_tree.R
\name{evolution_tree}
\alias{evolution_tree}
\title{Evolution Tree}
\usage{
evolution_tree(name_val)
}
\arguments{
\item{name_val}{}
}
\value{

}
\description{
Get the evolution tree of a pokemon character in a data frame
Note: this version only creates evolution trees for babies
}
\examples{

}
