% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ev_id.R
\name{find_ev_id}
\alias{find_ev_id}
\title{Find evolution ID}
\usage{
find_ev_id(name_val)
}
\arguments{
\item{name_val}{}
}
\value{
id
}
\description{
Find evolution ID
}
\examples{
find_ev_id("bulbasaur")

}
