% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_berry.R
\name{get_berry}
\alias{get_berry}
\title{Return a data frame containing information about the specified berry.}
\usage{
get_berry(s)
}
\arguments{
\item{s}{A string specifying the berry name.}
}
\value{
A dataframe containing information about the specified berry.
}
\description{
Return a data frame containing information about the specified berry.
}
\examples{
data(berry_lookup)
get_berry("cheri")

}
