% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types.R
\name{get_types}
\alias{get_types}
\title{Get available types for given API parameter}
\usage{
get_types(type, limit = 10)
}
\arguments{
\item{type}{-- API query type of interest}

\item{limit}{-- the max. number of \code{type}'s to return}
}
\value{
The types that the API will recognize.
}
\description{
Get available types for given API parameter
}
\examples{
get_types(type = "generation")

}
