% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_exists.R
\name{pokemon_exists}
\alias{pokemon_exists}
\title{Test pokemon exists}
\usage{
pokemon_exists(name_val)
}
\arguments{
\item{name}{Name of pokemon}
}
\value{
Boolean TRUE if the pokemon exists in the API, FALSE otherwise
}
\description{
Tests whether the pokemon is available in the API by comparing to a stored data frame
}
