% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokedex.R
\name{search_pokedex}
\alias{search_pokedex}
\title{Use the Pokedex Type Search Function}
\usage{
search_pokedex(type1)
}
\arguments{
\item{type1}{A character string specifying the primary Pokémon type to search for (e.g., "Fire", "Water", "Electric").}
}
\value{
A list of character vectors containing the names of Pokémon whose \code{Type1} matches the input type.
}
\description{
The \code{search_pokedex()} function allows users to search for Pokémon based on their primary type (\code{Type1}), using a case-insensitive match. It returns a list of Pokémon names that match the specified type.
}
\examples{
search_pokedex("Fire")
search_pokedex("WaTer")
}
