% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{type_damage}
\alias{type_damage}
\title{DATASET_TITLE}
\format{A data frame with 324 rows and 3 variables:
\describe{
  \item{\code{identifier_damage}}{character the type of damage of the move}
  \item{\code{identifier_target}}{character the type of the pokemon being attacked}
  \item{\code{damage_factor}}{double mutliplier}
}}
\source{
\url{https://github.com/veekun/pokedex}
}
\usage{
type_damage
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
