% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_effectiveness.R
\name{type_effectiveness}
\alias{type_effectiveness}
\title{Discover which types are effective against one another}
\usage{
type_effectiveness(type, strength = "all", direction = "all")
}
\arguments{
\item{type}{-- type}

\item{strength}{-- strength}

\item{direction}{-- direction}
}
\value{
Printed statements
}
\description{
Discover which types are effective against one another
}
\examples{
type_effectiveness(type="fire", strength="strong", direction="against")
}
