% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution_tree.R
\name{evolution_tree}
\alias{evolution_tree}
\title{Evolution Tree}
\usage{
evolution_tree(name_val)
}
\arguments{
\item{name_val}{The name of a Pokemon character, a string}
}
\value{
A data frame containing the names of the characters the name_val evolves into, the trigger that causes it, and the minimum level it occurs at
}
\description{
Get the evolution tree of a pokemon character in a data frame
Note: this version only creates evolution trees for babies
}
\examples{
evolution_tree("bulbasaur")

}
