% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_exists.R
\name{pokemon_exists}
\alias{pokemon_exists}
\title{Test pokemon exists}
\usage{
pokemon_exists(name_val)
}
\arguments{
\item{name_val}{Name of pokemon, a string}
}
\value{
Boolean TRUE if the pokemon exists in the evolution-chain API as a baby, FALSE otherwise
}
\description{
Tests whether the pokemon is a baby available in the evolution-chain API by comparing to a stored data frame.
}
\details{
This is intended to be used internally.
}
