% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_photo.R
\name{pokemon_photo}
\alias{pokemon_photo}
\title{Plot specified Pokemon's PNG image}
\usage{
pokemon_photo(name)
}
\arguments{
\item{name}{A character object that is the name of the Pokemon}
}
\value{
A rastergrob. See "Details" section of \code{\link[grid:grid.grob]{grid::grob()}} for more information.
}
\description{
\code{pokemon_photo()} displays a PNG image of a specified Pokemon (excluding variations in different Forms, Mega Evolutions, Region Forms) as a plot in a graphics device (e.g., the RStudio plot window).
}
\examples{
pokemon_photo("Pikachu")
}
