% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokedex.R
\name{search_pokedex}
\alias{search_pokedex}
\title{Use the Pokedex Type Search Function}
\usage{
search_pokedex(type_input, pokedex_dataset)
}
\description{
Given a comic number, the \code{pokedex()} function calls the Pokemon data set and returns the name of all of the Pokemon's that have Type 1 features that the user specified in the form of a list object.
}
